requirements.txt
discord.py>=1.4.1
gspread~=3.7.0
oauth2client~=4.1.3
psycopg2~=2.8.6